+++ dix/gc_priv.h
@@ -33,10 +33,10 @@ int SetDashes(GCPtr pGC, unsigned offset, unsigned ndash, unsigned char *pdash);
 int VerifyRectOrder(int nrects, xRectangle *prects, int ordering);
 
 int SetClipRects(GCPtr pGC,
-                int xOrigin,
-                int yOrigin,
-                int nrects,
+                INT16 xOrigin,
+                INT16 yOrigin,
+                size_t nrects,
                 xRectangle *prects,
-                int ordering);
+                BYTE ordering);
 
 #endif /* _XSERVER_DIX_GC_PRIV_H */
